+++ disassemblers/ofrak_capstone/CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased](https://github.com/redballoonsecurity/ofrak/tree/master)
 
+### Fixed
+- Capstone tests fixed to use `run_instruction_unpacker_test_case` ([#503](https://github.com/redballoonsecurity/ofrak/pull/503))
+
 ### Changed
 - Update to captione==5.0.0.post1.
+- Minor update to OFRAK Community License, add OFRAK Pro License ([#478](https://github.com/redballoonsecurity/ofrak/pull/478))
 
 ## 1.0.0 - 2022-01-25
 ### Added